// Admin/src/models/OrderRider.Model.ts
export interface IOrderRiderModel {
    Id: any
    PersonId: any
    ServiceTypeId: any
    ServiceType: any
    ModelVariantId: any
    ModelVariantName: any
    ThumbnailPath: any
    ProductTypeName: any
    BrandMasterName: any
    SeriesModelName: any
    UserMobile: any
    StartTime: any
    EndTime: any
    Address: any
    AppointmentCity: any
    AppointmentPincode: any
    UserName: any
    AppointmentDate: any
    OrderCode: any
    StatusId: any
    AssigneeId: any
    AssigneeName: any
    RowOrder: any
    TotalRowsCount: any
    StatusName: any
    SuggestedCost: any
    FinalPaid: any
    ReferralAmount: number,
    ReferralCodeId: any,
    TotalAmount: number,
    Amount: number,
    Adjustment?:number
}